+++ advanced_template/views/main_menu_view.py
@@ -1,4 +1,4 @@
-from advanced_template.views import BaseView
+from advanced_template.views.base_view import BaseView
 from advanced_template.special_scripts import BasicButton
 import arcade
 import arcade.gui
@@ -6,18 +6,15 @@ import arcade.gui
 
 class MainMenuView(BaseView):
     def __init__(self):
+        """
+        This is the view for the main menu.
+        """
         super().__init__(view_name="main_menu")
-
-    def on_show_view(self):
-        super().on_show_view()
 
     def setup(self):
         super().setup()
         arcade.set_background_color(arcade.color.BLUEBERRY)  # Arcade uses color not colour!)
-        self.setup_buttons()
 
-    def setup_buttons(self):
-        self.ui_manager = arcade.gui.UIManager()
         self.vertical_box = arcade.gui.UIBoxLayout()
 
         # Start button
@@ -25,14 +22,16 @@ class MainMenuView(BaseView):
 
         @start_button.event("on_click")
         def on_click_start(event):
-            print("STARTING!")
+            self.window.view_list["game"].setup()
+            self.window.show_view(self.window.view_list["game"])
 
+        # Settings Button
         settings_button = BasicButton(name="Settings Button", text="Settings")
 
         @settings_button.event("on_click")
         def on_click_settings(event):
-            print("Settings!")
-
+            self.window.view_list["settings"].setup()
+            self.window.show_view(self.window.view_list["settings"])
 
         # Quit button
         quit_button = BasicButton(name="Quit Button", text="Leave")
@@ -48,10 +47,16 @@ class MainMenuView(BaseView):
             arcade.gui.UIAnchorWidget(anchor_x="center_x", anchor_y="center_y", child=self.vertical_box))
         self.ui_manager.add(self.vertical_box)
 
+    def on_show_view(self):
+        super().on_show_view()
+
+    def on_hide_view(self):
+        super().on_hide_view()
+
     def on_draw(self):
         arcade.start_render()
-
         self.clear()
+        self.ui_manager.draw()
 
         arcade.draw_text(
             "Advanced Arcade Template",
@@ -63,6 +68,3 @@ class MainMenuView(BaseView):
             anchor_y="center",
         )
         arcade.draw_point(500, 500, arcade.color.PEAR, 20)
-
-        if self.ui_manager is not None:
-            self.ui_manager.draw()